refactor(core): move key management interfaces to Core
